+++ api/apps/api/src/modules/cost-surface/adapters/cost-surface-adapters.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ApiEventsModule } from '../../../api-events';
+import { ApiEventsModule } from '../../api-events';
 import { CostSurfaceEventsPort } from '../ports/cost-surface-events.port';
 import { CostSurfaceApiEvents } from './cost-surface-api-events';
