--- api/apps/api/src/modules/cost-surface/cost-surface-module-first.class.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { CostSurface } from './cost-surface.api.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([CostSurface])],
-})
-export class CostSurfaceModuleFirstClass {}
