+++ api/apps/api/src/modules/cost-surface/infra/delete-scenario.command.ts
