+++ api/apps/api/src/modules/cost-surface/infra/delete-scenario.handler.ts
@@ -1,7 +1,7 @@
 import { CommandHandler, IInferredCommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Scenario } from '../../scenario.api.entity';
+import { Scenario } from '../../scenarios/scenario.api.entity';
 import { DeleteScenario } from './delete-scenario.command';
 
 @CommandHandler(DeleteScenario)
