--- api/apps/api/src/modules/scenarios/cost-surface/cost-surface.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { CostSurfaceApplicationModule } from './application/cost-surface-application.module';
-
-@Module({
-  imports: [CostSurfaceApplicationModule],
-  providers: [],
-  exports: [],
-})
-export class CostSurfaceModule {}
