+++ api/apps/geoprocessing/src/modules/cost-surface/application/cost-surface-worker.service.ts
@@ -2,20 +2,20 @@ import { Injectable } from '@nestjs/common';
 import { EventBus } from '@nestjs/cqrs';
 import { Worker } from 'bullmq';
 
-import { JobInput, surfaceCostQueueName } from '@marxan/artifact-cache';
+import { JobInput, costSurfaceQueueName } from '@marxan/artifact-cache';
 import { WorkerBuilder } from '@marxan-geoprocessing/modules/worker';
 
-import { SurfaceCostProcessor } from './surface-cost-processor';
+import { CostSurfaceProcessor } from './cost-surface-processor.service';
 
 @Injectable()
-export class SurfaceCostWorker {
+export class CostSurfaceWorker {
   #worker: Worker;
 
   constructor(
     private readonly wrapper: WorkerBuilder,
-    private readonly processor: SurfaceCostProcessor,
+    private readonly processor: CostSurfaceProcessor,
     private readonly eventBus: EventBus,
   ) {
-    this.#worker = wrapper.build(surfaceCostQueueName, processor);
+    this.#worker = wrapper.build(costSurfaceQueueName, processor);
   }
 }
