+++ api/apps/geoprocessing/test/integration/cost-surface/cost-surface-job/process-shapefile-with-cost-surface.e2e-spec.ts
@@ -1,17 +1,17 @@
-import { SurfaceCostProcessor } from '@marxan-geoprocessing/modules/cost-surface/application/surface-cost-processor';
+import { CostSurfaceProcessor } from '@marxan-geoprocessing/modules/cost-surface/application/cost-surface-processor.service';
 import { INestApplication } from '@nestjs/common';
 import { PromiseType } from 'utility-types';
 import { bootstrapApplication, delay } from '../../../utils';
 import { createWorld } from './steps/world';
 
 let app: INestApplication;
-let sut: SurfaceCostProcessor;
+let sut: CostSurfaceProcessor;
 let world: PromiseType<ReturnType<typeof createWorld>>;
 
 describe(`given scenario has some planning units`, () => {
   beforeAll(async () => {
     app = await bootstrapApplication();
-    sut = app.get(SurfaceCostProcessor);
+    sut = app.get(CostSurfaceProcessor);
   });
   beforeEach(async () => {
     world = await createWorld(app);
